requirements.txt
flask>=3.0.0
pdfplumber>=0.10.3
pandas>=2.2.0
openpyxl>=3.1.2
gspread>=6.0.0
google-auth>=2.29.0
